Remove Repeating Elements
Add local limiter so menus do not repeat elements if there are less elements than the selected limit option. If the menu limit is 7 but there are only 4 elements, only 3 will elements at a time. Also, if less than 3 elements are in a menu, the limit will be the element length, so menus with 2 elements will display both elements, only once.